Update dropbox to v2.0 API
This change required the removal of the pop-up authentication and moves
to the redirect authentication.  This means the app should be connected
to dropbox *before* entering any data, otherwise the data will be wiped
after the redirect.  But after the redirect, the page can be bookmarked
as connected, so the user only needs to connect once.